% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tbl}
\alias{as_tbl}
\title{Convert object to tibble}
\usage{
as_tbl(x, ...)
}
\arguments{
\item{x}{A vector, list, or matrix. If a list, each element must have the same length.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tibble, class \code{c("tbl_df", "tbl", "data.frame")}
}
\description{
Convert object to tibble
}
